services/runpod-worker/requirements.txt
redis==5.0.1
supabase==2.3.0
numpy==1.26.4
scipy==1.12.0
python-dotenv==1.0.1
httpx==0.27.0
fastapi
uvicorn
runpod
requests
